refactor: change CreditStatus to Status as general component
