--- src/Nowruz/modules/credentials/creditStatus/creditStatus.module.scss
@@ -1,4 +1,0 @@
-.status {
-  border: 1px solid #d0d5dd;
-  @apply rounded-2xl flex items-center justify-center py-1 px-1 text-xs font-medium;
-}
