+++ src/Nowruz/modules/general/components/Status/index.types.ts
@@ -1,5 +1,5 @@
-export interface statusType {
-  icon: string;
+export interface StatusProps {
+  icon?: string;
   label: string;
   theme: 'primary' | 'secondary' | 'grey_blue' | 'error' | 'warning' | 'success';
 }
