+++ src/components/_exports.module.scss
@@ -62,6 +62,7 @@
     color_warning_50: $color-warning-50;
     color_warning_100: $color-warning-100;
     color_warning_300: $color-warning-300;
+    color_warning_500: $color-warning-500;
     color_warning_600: $color-warning-600;
     color_warning_700: $color-warning-700;
